refactor: 서술형 답변 유효성 검사 훅 구조 변경
